refactor(models/receipt.py): update total_price data type

+++ shopify/app/models/receipt.py
@@ -10,5 +10,5 @@ class Receipt(Base):
     date = Column(String(10))
     client_name = Column(String(40))
     client_email = Column(String(30))
-    total_price = Column(Integer)
+    total_price = Column(Double)
